# Conflicts:
#	sormas-backend/src/main/java/de/symeda/sormas/backend/externalsurveillancetool/ExternalSurveillanceToolGatewayFacadeEjb.java
#	sormas-ui/src/main/java/de/symeda/sormas/ui/externalsurveillanceservice/ExternalSurveillanceShareComponent.java

+++ sormas-ui/src/main/java/de/symeda/sormas/ui/externalmessage/ExternalMessageController.java
@@ -18,6 +18,7 @@ import static de.symeda.sormas.ui.externalmessage.processing.ExternalMessageProc
 
 import java.util.Collection;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.function.BiFunction;
 import java.util.stream.Collectors;
@@ -25,6 +26,7 @@ import java.util.stream.Collectors;
 import javax.naming.NamingException;
 
 import org.apache.commons.lang3.BooleanUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -44,23 +46,34 @@ import com.vaadin.ui.themes.ValoTheme;
 import de.symeda.sormas.api.FacadeProvider;
 import de.symeda.sormas.api.caze.CaseDataDto;
 import de.symeda.sormas.api.caze.CaseReferenceDto;
+import de.symeda.sormas.api.caze.surveillancereport.ReportingType;
+import de.symeda.sormas.api.caze.surveillancereport.SurveillanceReportDto;
 import de.symeda.sormas.api.externalmessage.ExternalMessageDto;
 import de.symeda.sormas.api.externalmessage.ExternalMessageIndexDto;
 import de.symeda.sormas.api.externalmessage.ExternalMessageResult;
 import de.symeda.sormas.api.externalmessage.ExternalMessageStatus;
+import de.symeda.sormas.api.externalmessage.ExternalMessageType;
+import de.symeda.sormas.api.externalmessage.labmessage.SampleReportDto;
 import de.symeda.sormas.api.i18n.Captions;
 import de.symeda.sormas.api.i18n.I18nProperties;
 import de.symeda.sormas.api.i18n.Strings;
 import de.symeda.sormas.api.i18n.Validations;
-import de.symeda.sormas.api.sample.SampleReferenceDto;
+import de.symeda.sormas.api.infrastructure.district.DistrictReferenceDto;
+import de.symeda.sormas.api.infrastructure.facility.FacilityDto;
+import de.symeda.sormas.api.infrastructure.facility.FacilityReferenceDto;
+import de.symeda.sormas.api.infrastructure.facility.FacilityType;
+import de.symeda.sormas.api.infrastructure.region.RegionReferenceDto;
+import de.symeda.sormas.api.sample.SampleDto;
 import de.symeda.sormas.api.user.UserReferenceDto;
+import de.symeda.sormas.api.user.UserRight;
+import de.symeda.sormas.api.utils.DataHelper;
 import de.symeda.sormas.ui.ControllerProvider;
 import de.symeda.sormas.ui.SormasUI;
 import de.symeda.sormas.ui.UserProvider;
 import de.symeda.sormas.ui.externalmessage.labmessage.LabMessageProcessingFlow;
 import de.symeda.sormas.ui.externalmessage.labmessage.LabMessageSlider;
 import de.symeda.sormas.ui.externalmessage.labmessage.RelatedLabMessageHandler;
-import de.symeda.sormas.ui.externalmessage.labmessage.processing.SampleAndPathogenTests;
+import de.symeda.sormas.ui.externalmessage.labmessage.processing.RelatedSamplesReportsAndPathogenTests;
 import de.symeda.sormas.ui.externalmessage.physiciansreport.PhysiciansReportProcessingFlow;
 import de.symeda.sormas.ui.externalmessage.processing.flow.ProcessingResult;
 import de.symeda.sormas.ui.externalmessage.processing.flow.ProcessingResultStatus;
@@ -78,9 +91,9 @@ public class ExternalMessageController {
 		relatedLabMessageHandler = new RelatedLabMessageHandler();
 	}
 
-	public void showExternalMessage(String labMessageUuid, boolean withActions, Runnable onFormActionPerformed) {
+	public void showExternalMessage(String MessageUuid, boolean withActions, Runnable onFormActionPerformed) {
 
-		ExternalMessageDto newDto = FacadeProvider.getExternalMessageFacade().getByUuid(labMessageUuid);
+		ExternalMessageDto newDto = FacadeProvider.getExternalMessageFacade().getByUuid(MessageUuid);
 		VerticalLayout layout = new VerticalLayout();
 		layout.setMargin(true);
 
@@ -112,7 +125,7 @@ public class ExternalMessageController {
 		LabMessageProcessingFlow flow = new LabMessageProcessingFlow();
 
 		flow.run(labMessage, relatedLabMessageHandler)
-			.handle((BiFunction<? super ProcessingResult<SampleAndPathogenTests>, Throwable, Void>) (result, exception) -> {
+			.handle((BiFunction<? super ProcessingResult<RelatedSamplesReportsAndPathogenTests>, Throwable, Void>) (result, exception) -> {
 				if (exception != null) {
 					logger.error("Unexpected exception while processing lab message", exception);
 
@@ -129,7 +142,15 @@ public class ExternalMessageController {
 				if (status == ProcessingResultStatus.CANCELED_WITH_CORRECTIONS) {
 					showCorrectionsSavedPopup();
 				} else if (status == ProcessingResultStatus.DONE) {
-					markExternalMessageAsProcessed(labMessage, result.getData().getSample().toReference());
+
+					CaseReferenceDto caseReference =
+						result.getData().getRelatedSampleReportsWithSamples().values().stream().findFirst().get().getAssociatedCase();
+					SurveillanceReportDto surveillanceReport = null;
+					if (caseReference != null) {
+						surveillanceReport = createSurveillanceReport(labMessage, caseReference);
+						FacadeProvider.getSurveillanceReportFacade().save(surveillanceReport);
+					}
+					markExternalMessageAsProcessed(labMessage, result.getData().getRelatedSampleReportsWithSamples(), surveillanceReport);
 					SormasUI.get().getNavigator().navigateTo(ExternalMessagesView.VIEW_NAME);
 				}
 
@@ -156,7 +177,10 @@ public class ExternalMessageController {
 			ProcessingResultStatus status = result.getStatus();
 
 			if (status == ProcessingResultStatus.DONE) {
-				markExternalMessageAsProcessed(labMessage, result.getData().toReference());
+				CaseReferenceDto caseReferenceDto = result.getData().toReference();
+				SurveillanceReportDto surveillanceReport = createSurveillanceReport(labMessage, caseReferenceDto);
+				FacadeProvider.getSurveillanceReportFacade().save(surveillanceReport);
+				markExternalMessageAsProcessed(labMessage, surveillanceReport);
 				SormasUI.get().getNavigator().navigateTo(ExternalMessagesView.VIEW_NAME);
 			}
 
@@ -164,16 +188,78 @@ public class ExternalMessageController {
 		});
 	}
 
-	public void markExternalMessageAsProcessed(ExternalMessageDto externalMessage, SampleReferenceDto sample) {
-		externalMessage.setSample(sample);
+	public void markExternalMessageAsProcessed(
+		ExternalMessageDto externalMessage,
+		Map<SampleReportDto, SampleDto> relatedSampleReports,
+		SurveillanceReportDto surveillanceReport) {
+
+		for (Map.Entry<SampleReportDto, SampleDto> entry : relatedSampleReports.entrySet()) {
+			entry.getKey().setSample(entry.getValue().toReference());
+		}
+		if (surveillanceReport != null) {
+			externalMessage.setSurveillanceReport(surveillanceReport.toReference());
+		}
 		externalMessage.setStatus(ExternalMessageStatus.PROCESSED);
 		FacadeProvider.getExternalMessageFacade().save(externalMessage);
 	}
 
-	public void markExternalMessageAsProcessed(ExternalMessageDto externalMessage, CaseReferenceDto caze) {
-		externalMessage.setCaze(caze);
+	public void markExternalMessageAsProcessed(ExternalMessageDto externalMessage, SurveillanceReportDto surveillanceReport) {
+		externalMessage.setSurveillanceReport(surveillanceReport.toReference());
 		externalMessage.setStatus(ExternalMessageStatus.PROCESSED);
 		FacadeProvider.getExternalMessageFacade().save(externalMessage);
+	}
+
+	protected SurveillanceReportDto createSurveillanceReport(ExternalMessageDto externalMessage, CaseReferenceDto caze) {
+		SurveillanceReportDto surveillanceReport = SurveillanceReportDto.build(caze, FacadeProvider.getUserFacade().getCurrentUserAsReference());
+		setSurvReportFacility(surveillanceReport, externalMessage, caze);
+		surveillanceReport.setReportDate(externalMessage.getMessageDateTime());
+		surveillanceReport.setExternalId(externalMessage.getReportMessageId());
+		setSurvReportingType(surveillanceReport, externalMessage);
+		return surveillanceReport;
+	}
+
+	private void setSurvReportFacility(SurveillanceReportDto surveillanceReport, ExternalMessageDto externalMessage, CaseReferenceDto caze) {
+		FacilityReferenceDto reporterReference = ExternalMessageMapper.getFacilityReference(externalMessage.getReporterExternalIds());
+		FacilityDto reporter;
+		if (reporterReference != null) {
+			reporter = FacadeProvider.getFacilityFacade().getByUuid(reporterReference.getUuid());
+			surveillanceReport.setFacility(reporterReference);
+			if (FacilityDto.OTHER_FACILITY_UUID.equals(reporter.getUuid())) {
+				surveillanceReport.setFacilityDetails(I18nProperties.getCaption(Captions.unknown));
+			}
+			surveillanceReport.setFacilityDistrict(reporter.getDistrict());
+			surveillanceReport.setFacilityRegion(reporter.getRegion());
+			surveillanceReport.setFacilityType(reporter.getType());
+		} else {
+			reporter = FacadeProvider.getFacilityFacade().getByUuid(FacilityDto.OTHER_FACILITY_UUID);
+			surveillanceReport.setFacility(reporter != null ? reporter.toReference() : null);
+			String reporterName = externalMessage.getReporterName();
+			if (StringUtils.isNotBlank(reporterName)) {
+				surveillanceReport.setFacilityDetails(reporterName);
+			} else {
+				surveillanceReport.setFacilityDetails(I18nProperties.getCaption(Captions.unknown));
+			}
+			DataHelper.Pair<RegionReferenceDto, DistrictReferenceDto> regionAndDistrict =
+				FacadeProvider.getCaseFacade().getRegionAndDistrictRefsOf(caze);
+			surveillanceReport.setFacilityRegion(regionAndDistrict.getElement0());
+			surveillanceReport.setFacilityDistrict(regionAndDistrict.getElement1());
+			if (ExternalMessageType.LAB_MESSAGE.equals(externalMessage.getType())) {
+				surveillanceReport.setFacilityType(FacilityType.LABORATORY);
+			} else if (ExternalMessageType.PHYSICIANS_REPORT.equals(externalMessage.getType())) {
+				surveillanceReport.setFacilityType(FacilityType.HOSPITAL);
+			}
+		}
+	}
+
+	protected void setSurvReportingType(SurveillanceReportDto surveillanceReport, ExternalMessageDto externalMessage) {
+		if (ExternalMessageType.LAB_MESSAGE.equals(externalMessage.getType())) {
+			surveillanceReport.setReportingType(ReportingType.LABORATORY);
+		} else if (ExternalMessageType.PHYSICIANS_REPORT.equals(externalMessage.getType())) {
+			surveillanceReport.setReportingType(ReportingType.DOCTOR);
+		} else {
+			throw new UnsupportedOperationException(
+				String.format("There is no reporting type defined for this type of external message: %s", externalMessage.getType()));
+		}
 	}
 
 	public void assignAllSelectedItems(Collection<ExternalMessageIndexDto> selectedRows, Runnable callback) {
@@ -223,23 +309,25 @@ public class ExternalMessageController {
 		buttonsPanel.setMargin(false);
 		buttonsPanel.setSpacing(true);
 
-		Button deleteButton = ButtonHelper.createButton(
-			Captions.actionDelete,
-			I18nProperties.getCaption(Captions.actionDelete),
-			e -> VaadinUiUtil.showDeleteConfirmationWindow(
-				String.format(I18nProperties.getString(Strings.confirmationDeleteEntity), I18nProperties.getCaption(Captions.ExternalMessage)),
-				() -> {
-					if (FacadeProvider.getExternalMessageFacade().isProcessed(externalMessage.getUuid())) {
-						showAlreadyProcessedPopup(null, false);
-					} else {
-						FacadeProvider.getExternalMessageFacade().deleteExternalMessage(externalMessage.getUuid());
-						callback.run();
-					}
-				}),
-			ValoTheme.BUTTON_DANGER,
-			CssStyles.BUTTON_BORDER_NEUTRAL);
+		if (UserProvider.getCurrent().hasUserRight(UserRight.EXTERNAL_MESSAGE_DELETE)) {
+			Button deleteButton = ButtonHelper.createButton(
+				Captions.actionDelete,
+				I18nProperties.getCaption(Captions.actionDelete),
+				e -> VaadinUiUtil.showDeleteConfirmationWindow(
+					String.format(I18nProperties.getString(Strings.confirmationDeleteEntity), I18nProperties.getCaption(Captions.ExternalMessage)),
+					() -> {
+						if (FacadeProvider.getExternalMessageFacade().isProcessed(externalMessage.getUuid())) {
+							showAlreadyProcessedPopup(null, false);
+						} else {
+							FacadeProvider.getExternalMessageFacade().deleteExternalMessage(externalMessage.getUuid());
+							callback.run();
+						}
+					}),
+				ValoTheme.BUTTON_DANGER,
+				CssStyles.BUTTON_BORDER_NEUTRAL);
 
-		buttonsPanel.addComponent(deleteButton);
+			buttonsPanel.addComponent(deleteButton);
+		}
 
 		Button unclearButton = ButtonHelper.createButton(
 			Captions.actionUnclearLabMessage,
@@ -291,7 +379,7 @@ public class ExternalMessageController {
 			Button shareButton = ButtonHelper.createIconButton(
 				Captions.sormasToSormasSendLabMessage,
 				VaadinIcons.SHARE,
-				e -> ControllerProvider.getSormasToSormasController().shareLabMessage(externalMessage, callback));
+				e -> ControllerProvider.getSormasToSormasController().shareExternalMessage(externalMessage, callback));
 
 			buttonsPanel.addComponent(shareButton);
 		}
